Fix windows build. Review URL: http://codereview.chromium.org/8725029
git-svn-id: http://v8.googlecode.com/svn/branches/bleeding_edge@10086 ce2b1a6d-e550-0410-aec6-3dcde31c8c00

+++ src/mark-compact.cc
@@ -424,7 +424,7 @@ void MarkCompactCollector::CollectEvacuationCandidates(PagedSpace* space) {
   const int kMaxMaxEvacuationCandidates = 1000;
   int max_evacuation_candidates = Min(
     kMaxMaxEvacuationCandidates,
-    static_cast<int>(sqrt(number_of_pages / 2) + 1));
+    static_cast<int>(sqrt(static_cast<double>(number_of_pages / 2)) + 1));
 
   if (FLAG_stress_compaction || FLAG_always_compact) {
     max_evacuation_candidates = kMaxMaxEvacuationCandidates;
